Allow config for custom ISL commit message template
# Summary
Regarding https://github.com/facebook/sapling/issues/632, allows a sapling config to be set which gives a default template for ISL title/description.

Currently namespaced under `committemplate.isl` (with other committemplate configs), but could flip this to `isl.committemplate` (with other isl configs).

Also could see this being set by a version controlled file in the repo to standardize by repo instead of sapling config (ex. `<root>/.sapling-isl-committemplate`), open to thoughts on that

This is necessary because debugcommitmessage isn't a public command, so this could potentially be fixed by adding that to sapling CLI instead.

# Test Plan
Templated this commit using it, after running `sl config -l committemplate.isl "<title># Summary\n\n# Test Plan\n\n"`
